create RestApi for create and update category of product

+++ product/src/main/java/net/oussama/product/DTO/ProductDto.java
@@ -15,5 +15,5 @@ public class ProductDto {
     private String description;
     private String price;
     private Integer quantity;
-    private Integer category_id;
+    private Long category_id;
 }
